+++ src/main/java/atlas/commands/Command.java
@@ -1,7 +1,7 @@
 package atlas.commands;
 
 import atlas.storage.Storage;
-import atlas.tasks.ItemList;
+import atlas.items.ItemList;
 import atlas.ui.Ui;
 
 public interface Command {
