+++ src/main/java/atlas/commands/ListCommand.java
@@ -1,10 +1,20 @@
 package atlas.commands;
 
 import atlas.storage.Storage;
-import atlas.tasks.ItemList;
+import atlas.items.ItemList;
 import atlas.ui.Ui;
 
+/**
+ * Represents a command that prints out all the lists
+ */
 public class ListCommand implements Command {
+
+    /**
+     * Executes the list command.
+     * @param itemList the list of items to be displayed
+     * @param ui the user interface used to display messages
+     * @param storage the storage system used to save data to
+     */
     @Override
     public void execute(ItemList itemList, Ui ui, Storage storage) {
         ui.printList(itemList);
