+++ src/main/java/atlas/commands/NewTodoCommand.java
@@ -2,16 +2,26 @@ package atlas.commands;
 
 import atlas.exceptions.EmptyTaskNameException;
 import atlas.storage.Storage;
-import atlas.tasks.Item;
-import atlas.tasks.ItemList;
+import atlas.items.Item;
+import atlas.items.ItemList;
 import atlas.ui.Ui;
 
+/**
+ * Represents a command that creates a new todo
+ */
 public class NewTodoCommand implements Command {
     String input;
 
     public NewTodoCommand(String input) {
         this.input = input;
     }
+
+    /**
+     * Executes the new todo command.
+     * @param itemList the list of items to be updated
+     * @param ui the user interface used to display messages
+     * @param storage the storage system used to save data to
+     */
     @Override
     public void execute(ItemList itemList, Ui ui, Storage storage) {
         try {
