+++ src/main/java/atlas/exceptions/EmptyTaskNameException.java
@@ -1,5 +1,8 @@
 package atlas.exceptions;
 
+/**
+ * Represents an exception thrown when task name is empty
+ */
 public class EmptyTaskNameException extends Exception {
     public EmptyTaskNameException() {
         super("Task name cannot be empty.");
