+++ src/main/java/atlas/exceptions/InvalidFormatDeadlineException.java
@@ -1,5 +1,8 @@
 package atlas.exceptions;
 
+/**
+ * Represents an exception thrown for deadlines with the wrong format.
+ */
 public class InvalidFormatDeadlineException extends Exception {
     public InvalidFormatDeadlineException() {
         super("Format: deadline <name> /by <datetime>");
