+++ src/main/java/atlas/exceptions/UnknownCommandException.java
@@ -1,4 +1,7 @@
 package atlas.exceptions;
 
+/**
+ * Represents an exception thrown for unknown commands.
+ */
 public class UnknownCommandException extends Exception{
 }
