+++ src/test/java/atlas/items/ItemListTest.java
@@ -1,4 +1,4 @@
-package atlas.tasks;
+package atlas.items;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
